Fixing more things:
Cleaning pragma lines, useles with gcc.
Because building binutils warns at every step
miserably.

exec*() family re-declared with newest prototype.
POSIX is now newer and they were warning without reason.

Added some needed things more. Now we have things that linux
guesses are in headers like resources.h, and I did a few ifdefs
in binutils to avoid stupids non-standard headers includements.
Harvey has all of that functions in APEX, but in standard places,
not where GNU wants. For example bzero().

libbsd will be included into libap, once we have native toolchain
running, because bsd.h is a f* waste of time and a rare bridge
between string.h and strings.h. And now, with musl source imported,
all is polluted with BSD things, in 2016 all of them normal, usual
and mostly standard. We're not BL guys, so there's no reason to
differentiate one code from another. ap/stdio dir has many things
from BSD sources, that Unix systems were including during years.

+++ lib/bsd/strncasecmp.c
@@ -12,7 +12,7 @@
 typedef unsigned char uchar;
 
 int
-strncasecmp(const char *s1, const char *s2, int n)
+strncasecmp(const char *s1, const char *s2, size_t n)
 {
 	int c1, c2;
